+++ conda/robocorp.go
@@ -122,6 +122,7 @@ func CondaExecutionEnvironment(location string, inject []string, full bool) []st
 		"RCC_ENVIRONMENT_HASH="+common.EnvironmentHash,
 		"RCC_INSTALLATION_ID="+xviper.TrackingIdentity(),
 		"RCC_TRACKING_ALLOWED="+fmt.Sprintf("%v", xviper.CanTrack()),
+		"RCC_EXE="+common.BinRcc(),
 		"RCC_VERSION="+common.Version,
 		"TEMP="+common.RobocorpTemp(),
 		"TMP="+common.RobocorpTemp(),